pyproject.toml
[tool.poetry]
name = "franklin-the-undying"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyNaCl = "^1.4.0"
discord = "^1.7.3"
Flask = "^2.0.2"
gTTS = "^2.2.3"
python-dotenv = "^0.19.2"
waitress = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
